requirements.txt
# Core dependencies for conversational customer service agent
groq==0.9.0
assemblyai==0.21.0
langgraph==0.6.5
pydantic==2.11.7
python-dotenv==1.1.1

# Audio processing
sounddevice==0.4.7
numpy==1.26.4
pyaudio==0.2.11

# Text-to-Speech
pyttsx3==2.90

# Audio/multimedia support
pygame==2.5.2

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0

# Optional: For advanced audio processing
scipy>=1.10.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utility libraries
requests>=2.31.0
typing-extensions>=4.5.0